% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_action_button_animated.R
\name{action_button_animated}
\alias{action_button_animated}
\title{Animated action button}
\usage{
action_button_animated(
  inputId,
  visible_content,
  hidden_content,
  class = "basic fade",
  ...
)
}
\arguments{
\item{inputId}{The `input` slot that will be used to access the value}

\item{visible_content}{Content of button}

\item{hidden_content}{Content of button when mouse is on button}

\item{class}{Additional classes for the button}

\item{...}{Attributes to pass to button}
}
\description{
Creates an animated action button. It's value is initially zero
  and increments by one each time it's pressed.
}
