% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_longest_dist_obs.R
\name{get_longest_dist_obs}
\alias{get_longest_dist_obs}
\title{Longest distance observations}
\usage{
get_longest_dist_obs(data, ship_name)
}
\arguments{
\item{data}{The data that has ships observations}

\item{ship_name}{The name of the vessel}
}
\value{
A dataframe with two observations
}
\description{
Returns the two consecutive observations of a vessel with the longest sailed distance
}
\examples{
# The two consecutive observations describing the longest sailed distance for the vessel 'KAROLI'
get_longest_dist_obs(data = ships, ship_name = 'KAROLI')
}
