% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships.R
\name{haversine}
\alias{haversine}
\title{Calculate shortest distance between two points on earth}
\usage{
haversine(dt)
}
\arguments{
\item{dt}{a data.table}
}
\description{
Calculate shortest distance between two points on earth
}
