% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_longest_distance.R
\name{longest_distance}
\alias{longest_distance}
\title{Longest distance}
\usage{
longest_distance(data)
}
\arguments{
\item{data}{Data.frame with vessel data. Should contain at least columns:
\describe{
  \item{Lon}{Numeric, longitude}
  \item{Lat}{Numeric, latitude}
  \item{Datetime}{Datetime}
}}
}
\value{
Two row data frame with observation for which distance was the greatest
  and next consecutive row. `Dist` and `Type` columns are appended, first one
  yielding calculation of distance, second one indicating whether it is starting
  or ending point
}
\description{
Calculate longest distance between consecutive logs of ship's position
}
