% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships.R
\name{ships}
\alias{ships}
\title{R6 class for ships data}
\description{
R6 class for ships data

R6 class for ships data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ships$new()}}
\item \href{#method-display_ship}{\code{ships$display_ship()}}
\item \href{#method-clone}{\code{ships$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ships$new(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-display_ship"></a>}}
\if{latex}{\out{\hypertarget{method-display_ship}{}}}
\subsection{Method \code{display_ship()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ships$display_ship(vessel_type, shipID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ships$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
